api/requirements.txt
black>=23.11.0
docker>=6.1.3
fastapi>=0.68.0,<0.110.0
flake8>=6.1.0
google-analytics-data==0.18.7
google-api-core==2.18.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
isort>=5.12.0
polars>=0.20.23
pydantic>=1.8.0,<3.0.0
uvicorn>=0.15.0,<0.28.0
psutil>=5.9.6
redis>=5.0.1
requests>=2.31.0